+++ scripts/testing.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+pip3 install -r requirements.txt
